backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
langchain==0.0.354
langchain-google-genai==0.0.6
supabase==1.2.0
python-dotenv==0.21.1
pydantic==1.10.13
httpx==0.24.1
google-generativeai==0.3.2


